+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Routes - Expose the record services and the permissions middleware.
 
+### Changed
+- Technical - Rename RecordsExporter to ResourcesExporter in export.
+
 ## RELEASE 5.1.1 - 2019-11-18
 ### Changed
 - Technical - Change functions of filter parser to async.